requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=4.0,<5.0
djangorestframework-authtoken>=1.0,<2.0
